[KPSDK-23100][secure-subsystem_digest] fix build warning
- enumeration value is out of "int" range

+++ inc/fsl_sss_api.h
@@ -25,9 +25,9 @@
 typedef enum _sss_status
 {
     kStatus_SSS_Success = 0x5a5a5a5au,
-    kStatus_SSS_Fail = 0xc3c30000u,
-    kStatus_SSS_InvalidArgument = 0xc3c30001u,
-    kStatus_SSS_ResourceBusy = 0xc3c30002u,
+    kStatus_SSS_Fail = (int)0xc3c30000u,
+    kStatus_SSS_InvalidArgument = (int)0xc3c30001u,
+    kStatus_SSS_ResourceBusy = (int)0xc3c30002u,
 } sss_status_t;
 
 typedef enum _sss_type
